requirements.txt
# ─────────────────────────────────────────────
# Core Backend (FastAPI + Uvicorn)
# ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ─────────────────────────────────────────────
# Data models and settings
# ─────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings>=2.6.1
python-dotenv>=1.0.1
python-dateutil==2.9.0.post0
email-validator

# ─────────────────────────────────────────────
# HTTP clients
# ─────────────────────────────────────────────
requests>=2.32.3
httpx>=0.27

# ─────────────────────────────────────────────
# AI / OpenAI integration
# ─────────────────────────────────────────────
openai>=1.50.0

# ─────────────────────────────────────────────
# Visualization and PDF generation
# ─────────────────────────────────────────────
matplotlib==3.9.2
reportlab>=4.2.2

# ─────────────────────────────────────────────
# MongoDB (Ocult profiles)
# ─────────────────────────────────────────────
pymongo[srv]>=4.6
motor>=3.3.0
gunicorn>=21.2.0

# ─────────────────────────────────────────────
# Frontend (Streamlit UI)
# ─────────────────────────────────────────────
streamlit>=1.39.0
urllib3>=2.2.3     # required for Retry in Streamlit app

# ─────────────────────────────────────────────
# Development and Testing
# ─────────────────────────────────────────────
pytest==8.3.3

# ─────────────────────────────────────────────
# Optional: Postgres (remove if unused)
# ─────────────────────────────────────────────
psycopg[binary]
